First pass at object cache support. (Thank you Andy Skelton for writing batcache from where I borrowed some code and ideas)
git-svn-id: http://plugins.svn.wordpress.org/wp-super-cache/trunk@198399 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ wp-cache-phase2.php
@@ -230,21 +230,6 @@ function wp_cache_writers_exit() {
 	}
 }
 
-function get_current_url_supercache_dir() {
-	global $cached_direct_pages, $cache_path, $wp_cache_request_uri;
-	$uri = preg_replace('/[ <>\'\"\r\n\t\(\)]/', '', str_replace( '/index.php', '/', str_replace( '..', '', preg_replace("/(\?.*)?$/", '', $wp_cache_request_uri ) ) ) );
-	$uri = str_replace( '\\', '', $uri );
-	$dir = strtolower(preg_replace('/:.*$/', '',  $_SERVER["HTTP_HOST"])) . $uri; // To avoid XSS attacks
-	$dir = apply_filters( 'supercache_dir', $dir );
-	$dir = trailingslashit( $cache_path . 'supercache/' . $dir );
-	if( is_array( $cached_direct_pages ) && in_array( $_SERVER[ 'REQUEST_URI' ], $cached_direct_pages ) ) {
-		$dir = trailingslashit( ABSPATH . $uri );
-	}
-	$dir = str_replace( '//', '/', $dir );
-	if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "supercache dir: $dir", 5 );
-	return $dir;
-}
-
 function wp_cache_ob_callback( $buffer ) {
 	global $wp_cache_pages;
 	if( defined( 'DONOTCACHEPAGE' ) )
@@ -282,7 +267,7 @@ function wp_cache_get_ob(&$buffer) {
 	global $wp_cache_gzip_encoding, $super_cache_enabled, $cached_direct_pages;
 	global $wp_cache_404, $gzsize, $supercacheonly, $wp_cache_gzip_first, $wp_cache_gmt_offset;
 	global $blog_cache_dir, $wp_cache_request_uri, $wp_supercache_cache_list;
-	global $wp_cache_not_logged_in, $wp_cache_anon_only;
+	global $wp_cache_not_logged_in, $wp_cache_anon_only, $wp_cache_object_cache;
 
 	$new_cache = true;
 	$wp_cache_meta = '';
@@ -352,15 +337,23 @@ function wp_cache_get_ob(&$buffer) {
 		}
 	}
 
+	if ( $wp_cache_object_cache ) { // half on mode when using the object cache
+		$super_cache_enabled = false;
+		$supercacheonly = false;
+		wp_cache_init(); // PHP5 destroys objects during shutdown
+	}
+
 	if ( $cache_enabled ) {
 		// Open wp-cache cache file
 		if ( !$supercacheonly && ( !@file_exists( $blog_cache_dir . $cache_filename ) || ( @file_exists( $blog_cache_dir . $cache_filename ) && ( time() - @filemtime( $blog_cache_dir . $cache_filename ) ) > 5 ) ) ) {
-			$fr = @fopen($tmp_wpcache_filename, 'w');
-			if (!$fr) {
-				if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Error. Supercache could not write to " . str_replace( ABSPATH, '', $cache_path ) . $cache_filename, 1 );
-				$buffer .= "<!-- File not cached! Super Cache Couldn't write to: " . str_replace( ABSPATH, '', $cache_path ) . $cache_filename . " -->\n";
-				wp_cache_writers_exit();
-				return $buffer;
+			if ( false == $wp_cache_object_cache ) {
+				$fr = @fopen($tmp_wpcache_filename, 'w');
+				if (!$fr) {
+					if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Error. Supercache could not write to " . str_replace( ABSPATH, '', $cache_path ) . $cache_filename, 1 );
+					$buffer .= "<!-- File not cached! Super Cache Couldn't write to: " . str_replace( ABSPATH, '', $cache_path ) . $cache_filename . " -->\n";
+					wp_cache_writers_exit();
+					return $buffer;
+				}
 			}
 		}
 		$user_info = wp_cache_get_cookies_values();
@@ -404,6 +397,8 @@ function wp_cache_get_ob(&$buffer) {
 		$buffer .= "<!-- Live page served on " . gmdate('Y-m-d H:i:s', (time() + ( $wp_cache_gmt_offset * 3600)))  . " -->\n";
 	}
 
+	$added_cache = 0;
+	$oc_key = get_oc_key();
 	if ( preg_match('/<!--mclude|<!--mfunc/', $buffer)) { //Dynamic content
 		if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Dynamic content found in buffer.", 4 );
 		$store = preg_replace('|<!--mclude (.*?)-->(.*?)<!--/mclude-->|is', 
@@ -415,75 +410,86 @@ function wp_cache_get_ob(&$buffer) {
 		/* Clean function calls in tag */
 		$buffer = preg_replace('|<!--mclude (.*?)-->|is', '<!--mclude-->', $buffer);
 		$buffer = preg_replace('|<!--mfunc (.*?)-->|is', '<!--mfunc-->', $buffer);
-		if( $fr )
-			fputs($fr, $store);
-	} else {
-		$buffer = apply_filters( 'wpsupercache_buffer', $buffer );
-		if( $gz || $wp_cache_gzip_encoding ) {
-			if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Gzipping buffer.", 5 );
-			$gzdata = gzencode( $buffer . "<!-- Compression = gzip -->", 3, FORCE_GZIP );
+		if ( false == $wp_cache_object_cache ) {
+			if( $fr )
+				fputs($fr, $store);
+		} else {
+			wp_cache_set( $oc_key, $store, 'supercache', 30 ); // TODO set timelimit
+		}
 			$gzsize = strlen($gzdata);
 		}
 		if ($wp_cache_gzip_encoding) {
 			$wp_cache_meta[ 'headers' ][ 'Content-Encoding' ] = 'Content-Encoding: ' . $wp_cache_gzip_encoding;
 			$wp_cache_meta[ 'headers' ][ 'Vary' ] = 'Vary: Accept-Encoding, Cookie';
 			// Return uncompressed data & store compressed for later use
-			if( $fr ) {
-				if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Writing gzipped buffer to wp-cache cache file.", 5 );
-				fputs($fr, $gzdata);
+			if ( false == $wp_cache_object_cache ) {
+				if( $fr ) {
+					if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Writing gzipped buffer to wp-cache cache file.", 5 );
+					fputs($fr, $gzdata);
+				}
+			} else {
+				wp_cache_set( $oc_key . ".gz", $gzdata, 'supercache', 30 ); // TODO set timelimit
+				$added_cache = 1;
 			}
 		} else { // no compression
 			$wp_cache_meta[ 'headers' ][ 'Vary' ] = 'Vary: Cookie';
-			if( $fr ) {
-				if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Writing non-gzipped buffer to wp-cache cache file.", 5 );
-				fputs($fr, $buffer);
+			if ( false == $wp_cache_object_cache ) {
+				if( $fr ) {
+					if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Writing non-gzipped buffer to wp-cache cache file.", 5 );
+					fputs($fr, $buffer);
+				}
+			} else {
+				wp_cache_set( $oc_key, $buffer, 'supercache', 30 ); // TODO set timelimit
+				$added_cache = 1;
 			}
 		}
-		if( $fr2 ) {
-			if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Writing non-gzipped buffer to supercache file.", 5 );
-			fputs($fr2, $buffer . '<!-- super cache -->' );
+		if ( false == $wp_cache_object_cache ) {
+			if( $fr2 ) {
+				if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Writing non-gzipped buffer to supercache file.", 5 );
+				fputs($fr2, $buffer . '<!-- super cache -->' );
+			}
+			if( $gz ) {
+				if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Writing gzipped buffer to supercache file.", 5 );
+				fwrite($gz, $gzdata );
+			}
 		}
-		if( $gz ) {
-			if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Writing gzipped buffer to supercache file.", 5 );
-			fwrite($gz, $gzdata );
-		}
-		$buffer .= $log;
 	}
 	$new_cache = true;
-	$added_cache = 0;
-	if( $fr ) {
-		$supercacheonly = false;
-		fclose($fr);
-		if ( filesize( $tmp_wpcache_filename ) == 0 ) {
-			@unlink( $tmp_wpcache_filename );
-		} elseif ( !rename( $tmp_wpcache_filename, $blog_cache_dir . $cache_filename ) ) {
-			unlink( $blog_cache_dir . $cache_filename );
-			rename( $tmp_wpcache_filename, $blog_cache_dir . $cache_filename );
+	if ( false == $wp_cache_object_cache ) {
+		if( $fr ) {
+			$supercacheonly = false;
+			fclose($fr);
+			if ( filesize( $tmp_wpcache_filename ) == 0 ) {
+				@unlink( $tmp_wpcache_filename );
+			} elseif ( !rename( $tmp_wpcache_filename, $blog_cache_dir . $cache_filename ) ) {
+				unlink( $blog_cache_dir . $cache_filename );
+				rename( $tmp_wpcache_filename, $blog_cache_dir . $cache_filename );
+			}
+			if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Renamed temp wp-cache file to {$blog_cache_dir}$cache_filename", 5 );
+			$added_cache = 1;
 		}
-		if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Renamed temp wp-cache file to {$blog_cache_dir}$cache_filename", 5 );
-		$added_cache = 1;
-	}
-	if( $fr2 ) {
-		fclose($fr2);
-		if ( filesize( $tmp_cache_filename ) == 0 ) {
-			@unlink( $tmp_cache_filename );
-		} elseif ( !@rename( $tmp_cache_filename, $cache_fname ) ) {
-			@unlink( $cache_fname );
-			@rename( $tmp_cache_filename, $cache_fname );
+		if( $fr2 ) {
+			fclose($fr2);
+			if ( filesize( $tmp_cache_filename ) == 0 ) {
+				@unlink( $tmp_cache_filename );
+			} elseif ( !@rename( $tmp_cache_filename, $cache_fname ) ) {
+				@unlink( $cache_fname );
+				@rename( $tmp_cache_filename, $cache_fname );
+			}
+			if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Renamed temp supercache file to $cache_fname", 5 );
+			$added_cache = 1;
 		}
-		if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Renamed temp supercache file to $cache_fname", 5 );
-		$added_cache = 1;
-	}
-	if( $gz ) {
-		fclose($gz);
-		if ( filesize( $tmp_cache_filename . '.gz' ) == 0 ) {
-			@unlink( $tmp_cache_filename . '.gz' );
-		} elseif ( !@rename( $tmp_cache_filename . '.gz', $cache_fname . '.gz' ) ) {
-			@unlink( $cache_fname . '.gz' );
-			@rename( $tmp_cache_filename . '.gz', $cache_fname . '.gz' );
+		if( $gz ) {
+			fclose($gz);
+			if ( filesize( $tmp_cache_filename . '.gz' ) == 0 ) {
+				@unlink( $tmp_cache_filename . '.gz' );
+			} elseif ( !@rename( $tmp_cache_filename . '.gz', $cache_fname . '.gz' ) ) {
+				@unlink( $cache_fname . '.gz' );
+				@rename( $tmp_cache_filename . '.gz', $cache_fname . '.gz' );
+			}
+			if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Renamed temp supercache gz file to {$cache_fname}.gz", 5 );
+			$added_cache = 1;
 		}
-		if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Renamed temp supercache gz file to {$cache_fname}.gz", 5 );
-		$added_cache = 1;
 	}
 	if ( $added_cache && isset( $wp_supercache_cache_list ) && $wp_supercache_cache_list ) {
 		update_option( 'wpsupercache_count', ( get_option( 'wpsupercache_count' ) + 1 ) );
@@ -633,7 +639,7 @@ function wp_cache_phase2_clean_expired($file_prefix) {
 
 function wp_cache_shutdown_callback() {
 	global $cache_path, $cache_max_time, $file_expired, $file_prefix, $meta_file, $new_cache, $wp_cache_meta, $known_headers, $blog_id, $wp_cache_gzip_encoding, $gzsize, $cache_filename, $supercacheonly, $blog_cache_dir;
-	global $wp_cache_blog_charset, $wp_cache_request_uri, $wp_cache_key;
+	global $wp_cache_blog_charset, $wp_cache_request_uri, $wp_cache_key, $wp_cache_object_cache;
 
 	$wp_cache_meta[ 'uri' ] = $_SERVER["SERVER_NAME"].preg_replace('/[ <>\'\"\r\n\t\(\)]/', '', $wp_cache_request_uri); // To avoid XSS attacks
 	$wp_cache_meta[ 'blog_id' ] = $blog_id;
@@ -695,17 +701,24 @@ function wp_cache_shutdown_callback() {
 		$serial = serialize($wp_cache_meta);
 		if( wp_cache_writers_entry() ) {
 			if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Writing meta file: {$blog_cache_dir}meta/{$meta_file}", 2 );
-			$tmp_meta_filename = $blog_cache_dir . 'meta/' . uniqid( mt_rand(), true ) . '.tmp';
-			$fr = @fopen( $tmp_meta_filename, 'w');
-			if( !$fr )
-				@mkdir( $blog_cache_dir . 'meta' );
-			$fr = fopen( $tmp_meta_filename, 'w');
-			fputs($fr, $serial);
-			fclose($fr);
-			@chmod( $tmp_meta_filename, 0666 & ~umask());
-			if( !@rename( $tmp_meta_filename, $blog_cache_dir . 'meta/' . $meta_file ) ) {
-				unlink( $blog_cache_dir . 'meta/' . $meta_file );
-				rename( $tmp_meta_filename, $blog_cache_dir . 'meta/' . $meta_file );
+			if ( false == $wp_cache_object_cache ) {
+				$tmp_meta_filename = $blog_cache_dir . 'meta/' . uniqid( mt_rand(), true ) . '.tmp';
+				$fr = @fopen( $tmp_meta_filename, 'w');
+				if( !$fr )
+					@mkdir( $blog_cache_dir . 'meta' );
+				$fr = fopen( $tmp_meta_filename, 'w');
+				fputs($fr, $serial);
+				fclose($fr);
+				@chmod( $tmp_meta_filename, 0666 & ~umask());
+				if( !@rename( $tmp_meta_filename, $blog_cache_dir . 'meta/' . $meta_file ) ) {
+					unlink( $blog_cache_dir . 'meta/' . $meta_file );
+					rename( $tmp_meta_filename, $blog_cache_dir . 'meta/' . $meta_file );
+				}
+			} else {
+				$oc_key = get_oc_key() . ".meta";
+				if ( gzip_accepted() )
+					$oc_key .= ".gz";
+				wp_cache_set( $oc_key, $serial, 'supercache', 30 );
 			}
 			wp_cache_writers_exit();
 		}
@@ -762,17 +775,25 @@ function wp_cache_get_postid_from_comment( $comment_id, $status = 'NA' ) {
 
 /* Clear out the cache directory. */
 function wp_cache_clear_cache() {
-	global $cache_path;
-	prune_super_cache( $cache_path . 'supercache/', true );
-	prune_super_cache( $cache_path, true );
+	global $cache_path, $wp_cache_object_cache;
+	if ( $wp_cache_object_cache ) {
+		reset_oc_version();
+	} else {
+		prune_super_cache( $cache_path . 'supercache/', true );
+		prune_super_cache( $cache_path, true );
+	}
 }
 
 function wp_cache_post_edit($post_id) {
 	global $wp_cache_clear_on_post_edit, $cache_path, $blog_cache_dir;
 	if( $wp_cache_clear_on_post_edit ) {
 		if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Clearing cache $blog_cache_dir and {$cache_path}supercache/ on post edit per config.", 2 );
-		prune_super_cache( $blog_cache_dir, true );
-		prune_super_cache( $cache_path . 'supercache/', true );
+		if ( $wp_cache_object_cache ) {
+			reset_oc_version();
+		} else {
+			prune_super_cache( $blog_cache_dir, true );
+			prune_super_cache( $cache_path . 'supercache/', true );
+		}
 	} else {
 		if ( isset( $GLOBALS[ 'wp_super_cache_debug' ] ) && $GLOBALS[ 'wp_super_cache_debug' ] ) wp_cache_debug( "Clearing cache for post $post_id on post edit.", 2 );
 		wp_cache_post_change( $post_id );
@@ -780,8 +801,11 @@ function wp_cache_post_edit($post_id) {
 }
 
 function wp_cache_post_id_gc( $siteurl, $post_id ) {
-	global $cache_path;
+	global $cache_path, $wp_cache_object_cache;
 	
+	if ( $wp_cache_object_cache )
+		reset_oc_version();
+
 	$post_id = intval( $post_id );
 	if( $post_id == 0 )
 		return;
